libeos-updater-util/util: Fix potentially truncated signal ID
If this g_signal_connect() call ends up returning a very high signal ID
(unlikely, but possible), the most significant bits can be truncated on
64-bit architectures, leaving the signal ID invalid for that (object,
signal, user_data) combination.

Fix that by using the right type to store the signal ID.


https://phabricator.endlessm.com/T19388

+++ libeos-updater-util/util.c
@@ -373,7 +373,7 @@ struct _EosQuitFile
   GObject parent_instance;
 
   GFileMonitor *monitor;
-  guint signal_id;
+  gulong signal_id;
   guint timeout_seconds;
   guint timeout_id;
   EosQuitFileCheckCallback callback;
@@ -395,7 +395,7 @@ quit_clear_user_data (EosQuitFile *quit_file)
 static void
 quit_disconnect_monitor (EosQuitFile *quit_file)
 {
-  guint id = quit_file->signal_id;
+  gulong id = quit_file->signal_id;
 
   quit_file->signal_id = 0;
   if (id > 0)
